+++ src/main/java/domain/AttendanceBook.java
@@ -1,6 +1,6 @@
 package domain;
 
-import dto.AddAttendanceRequest;
+import dto.request.AddAttendanceRequest;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.ArrayList;
