+++ src/main/java/domain/timeprovider/NowTimeProvider.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.timeprovider;
 
 import java.time.Clock;
 import java.time.LocalDate;
