+++ src/main/java/domain/timeprovider/TimeProvider.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.timeprovider;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
