+++ src/main/java/dto/response/AttendanceRecordResponse.java
@@ -1,4 +1,4 @@
-package dto;
+package dto.response;
 
 public record AttendanceRecordResponse(
         String date,
