+++ src/main/java/dto/response/AttendanceStatusCountResponse.java
@@ -1,4 +1,4 @@
-package dto;
+package dto.response;
 
 public record AttendanceStatusCountResponse(
         int attendCount,
