+++ src/main/java/dto/response/CrewWithPenaltyResponse.java
@@ -1,4 +1,4 @@
-package dto;
+package dto.response;
 
 import domain.AttendanceStatus;
 import domain.Crew;
