+++ src/main/java/util/filereader/CsvFileReader.java
@@ -1,4 +1,4 @@
-package util;
+package util.filereader;
 
 import java.io.IOException;
 import java.nio.file.Files;
