+++ src/main/java/util/filereader/FileReader.java
@@ -1,4 +1,4 @@
-package util;
+package util.filereader;
 
 import java.util.List;
