+++ src/main/java/util/inputreader/InputReader.java
@@ -1,4 +1,4 @@
-package util;
+package util.inputreader;
 
 public interface InputReader {
     public String readline();
